refactor(auth): split OTP page into modular components
